requirements.txt
django
djangorestframework
Celery
django-celery-beat
sqlalchemy
